move functiosn to aws package
